feat: change field to auth and update patch API
