+++ modules/auth/auth.model.js
@@ -16,14 +16,11 @@
             type: String,
             required: true
         },
-        phoneNumber: {
-            type: Number,
+        password: {
+            type: String,
             required: true
         },
-        address: String,
-        city: String,
-        state: String,
-        zipCode: String,
+        phoneNumber: Number,
         country: String
     });
