+++ modules/auth/auth.module.js
@@ -1,0 +1,15 @@
+(function () {
+    'use strict';
+
+    module.exports = init;
+
+    function init() {
+        return {
+            AuthController: requirauth.e('./controller'),
+            AuthMiddleware: require('./auth.middleware'),
+            AuthService: require('./auth.service'),
+            AuthModel: require('./auth.model')
+        }
+    }
+
+})();
